refactor(ui): standardize research date range formatting
